.github/generator/requirements.txt
pip==23.1.2
PyYAML==6.0
setuptools==67.7.2
wheel==0.40.0
